server/requirements.txt
livekit-plugins-openai
livekit-plugins-silero
livekit-agents
python-dotenv
mcp>=1.0.0
fastmcp>=2.0.0
aiohttp>=3.9.0
fastapi
uvicorn[standard]
sse-starlette
pydantic
httpx
# PostgreSQL and SQLAlchemy dependencies
sqlalchemy[asyncio]>=2.0.0
asyncpg
alembic
psycopg2-binary
# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx  # Already listed above for testing FastAPI
